Avoid stomping on uboot.env during u-boot partition upgrade
New u-boot writes environment to `uboot.env` on FAT (instead of MMC).
It automatically finds the first FAT partition and uses that.
Unfortunately(?) this is our u-boot config partition, and the entire
contents (including `uboot.env`) was replaced on RAUC updates.

Mark the `boot` slot readonly, and no longer include the `boot` image
in the RAUC update bundle.  This means we no longer update the
partition automatically, which is almost a feature since it lets
people tweak their own uboot/devicetree config.  Will almost certainly
improve this in future, however.